refactor(LadderResult): 사다리 결과 값 package 조정
- LadderResult은 실제 Ladder 내부 동작에 영향이 없음으로 패키지 바깥으로 꺼냄
