+++ src/test/java/ladder/domain/ladder/LadderTest.java
@@ -1,6 +1,7 @@
 package ladder.domain.ladder;
 
 import ladder.domain.Gamers;
+import ladder.domain.LadderResult;
 import ladder.domain.dto.LadderResultDto;
 import ladder.domain.ladder.maker.MakeLadderStrategy;
 import org.junit.jupiter.api.BeforeEach;
